DSL simplification and class refactoring
Simplified the DSL by removing calculation from feature. Also refactored the classes and renamed their file names.
