+++ scripts/testing.py
@@ -5,7 +5,12 @@ import numpy as np
 root_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
 sys.path.append(root_dir)
 
-from mestDS.classes.ClimateHealth import MultipleClimateHealth
+from mestDS.classes.Simulation import Simulations
 
-ch_sim = MultipleClimateHealth("scripts/simulation5.yaml")
+ch_sim = Simulations("scripts/simulation5.yaml")
 ch_sim.simulate()
+ch_sim.convert_to_csvs("testing_minimalist_multiregion/")
+ch_sim.eval_chap_model("models/minimalist_multiregion")
+
+# ch_sim.convert_to_csvs("testing_minimalist_example_lag/")
+# ch_sim.eval_chap_model("models/minimalist_example_lag")
