+++ src/mestDS/__init__.py
@@ -2,5 +2,12 @@ from .simulate import (
     generate_data,
 )
 from .evaluate import evaluate_chap_model
-from .classes import RainSeason, Simulation, MultipleSimulations, Region
+from .classes import (
+    RainSeason,
+    SimulationLegacy,
+    MultipleSimulationsLegacy,
+    Region,
+    Simulation,
+    Simulations,
+)
 from .utils import train_test_split_csv, plot_data_with_sample_0
