+++ src/mestDS/classes/Feature.py
@@ -1,5 +1,3 @@
 class Feature:
     name: str
-    calculation: str | None
-    function: str | None
-    parameters: list[str]
+    function: str
