+++ testing_minimalist_multiregion/0/dataset_x_test.csv
@@ -1,0 +1,81 @@
+time_period,rainfall,mean_temperature,location
+0.11918459219279837,19.601666355205992,14.944425883044591,Finnmark
+0.19517200571725712,25.19594506858049,22.247263856857867,Finnmark
+0.3667322713762387,24.99561860603057,22.649117233487395,Finnmark
+0.4673444730897248,20.27702156806342,16.154521780921137,Finnmark
+0.6823396097127724,21.037555383872437,15.203619006378062,Finnmark
+0.6927907731359264,27.81718966027964,20.13171798636611,Finnmark
+0.7010283105185549,18.696245969897273,16.257342437047342,Finnmark
+0.7423361542111717,28.831987665742055,19.729266871901643,Finnmark
+0.7742298886105831,28.93589496747379,19.12299318081818,Finnmark
+0.7805727265336266,20.916073012558194,22.93160430172782,Finnmark
+0.8619971692004202,28.36868617503084,19.6659646819849,Finnmark
+0.9417656803763995,22.141205066625137,17.760135127740106,Finnmark
+0.9507722471710189,20.292529651596443,13.913375806879406,Finnmark
+0.9762185900065296,24.80383136966226,19.109201123631113,Finnmark
+0.9763827765624724,16.646496949328245,14.103176896055375,Finnmark
+1.0061822606478055,22.643022536638924,14.740644763448643,Finnmark
+1.0139730231195048,26.286570090287505,24.097532965751466,Finnmark
+1.0373265465253905,29.930809750478858,18.744436708434396,Finnmark
+1.0521538097296175,23.269415150277563,13.484537130506062,Finnmark
+1.0937288435924677,23.500229920911682,14.255474143096535,Finnmark
+0.16261640731057428,21.772592619588053,25,Troms
+0.3272979231388481,21.07023309254854,19.491026136814153,Troms
+0.4015178682815123,22.085085705973384,14.717434086823173,Troms
+0.46379089036405624,20.725259498132075,17.55666907754366,Troms
+0.47636947642974947,21.212295053407075,18.202075564794658,Troms
+0.5449471863560928,20.838374046875828,16.124026904753563,Troms
+0.7160920616405975,22.03410641712256,20.842984833645325,Troms
+0.7443111599469558,21.64318356843154,16.664469170757336,Troms
+0.744586883443375,27.936333229069408,17.815700723949284,Troms
+0.7446504990035601,17.31933536013001,14.767403146164678,Troms
+0.7530497301597526,26.32568860710812,26.214521302793816,Troms
+0.7942826198787734,29.591340798555507,18.508147636390657,Troms
+0.85542845325249,26.518062647884157,19.160137524122774,Troms
+0.9989424592275628,26.735277727861273,18.193067034450134,Troms
+1.0691066853011042,25.971239212703377,20.11624853137299,Troms
+1.0870340265402474,22.47981899311261,17.62483759970721,Troms
+1.0981796177254703,24.711817268344173,20.294843616383766,Troms
+1.1168133503514053,22.01082603372266,16.33508258443527,Troms
+1.1175170200631959,21.41437764823671,17.041076243209243,Troms
+1.2163210157908702,23.292336322608346,21.874032492794882,Troms
+0.15753632421154296,32.789011788602714,19.551256241452666,Nordland
+0.16457966770382934,19.672790588814546,16.881154601431888,Nordland
+0.24737160264135472,20.697193970894563,14.802818952469616,Nordland
+0.33815055037615555,29.18887310614576,19.842374998866763,Nordland
+0.3501512300176474,16.467282701264736,14.295269699610218,Nordland
+0.36660196683035823,27.77833186270155,19.4765019420638,Nordland
+0.37839435151502365,30.7230368467228,19.08748792846615,Nordland
+0.39470238184587836,31.141226969022117,19.89004290251875,Nordland
+0.39568517768248596,23.310219256624496,13.371121017404722,Nordland
+0.43756720720667525,26.507753139134927,19.27967709553165,Nordland
+0.4484852917407347,28.07252053092901,19.768015186481023,Nordland
+0.4586195310589325,30.14695898657695,19.561189543098934,Nordland
+0.49832623815897614,19.86047333454445,13.744134729871984,Nordland
+0.5281653937674378,25.13812593509959,16.20489331135978,Nordland
+0.5458690128724605,16.655336672437393,12.791113414080135,Nordland
+0.555163923538582,29.433681172802157,18.336279459062418,Nordland
+0.6192795376719142,19.695169624337808,14.68846032888861,Nordland
+0.6575427585179277,20.275003260047963,13.281886864418105,Nordland
+0.669088303052583,32.616362162264764,21.089094918620976,Nordland
+0.6757272612961174,28.668732254068466,17.074125903068833,Nordland
+0.07817073492641759,27.561399348706136,30.64534542305234,Trøndelag
+0.20795968851779276,27.403740258483175,25.16327468551458,Trøndelag
+0.2368690711721977,21.3712099136426,13.895654942605141,Trøndelag
+0.4550780725143516,24.847845807116553,25,Trøndelag
+0.48248558949374765,29.938401418702345,24.530156721786465,Trøndelag
+0.49078768121970173,19.90598492193794,13.326511209723726,Trøndelag
+0.5416875533737691,20.211803509054228,27.74294754003143,Trøndelag
+0.5566197123061257,18.50793369571012,14.09806251510403,Trøndelag
+0.5573285564903084,21.25022074218208,15.329459791782737,Trøndelag
+0.5582990440507196,18.807833787002565,13.859246541597509,Trøndelag
+0.6252029360680726,23.377716793345005,20.85947350962335,Trøndelag
+0.6649025817969281,17.511880137419666,15.524891611643739,Trøndelag
+0.7438852362911085,24.346921478807225,23.81695185425238,Trøndelag
+0.7774361499829475,19.497402085348757,14.094962552185436,Trøndelag
+0.84653265771923,25.66172040636693,19.60238359302318,Trøndelag
+0.8478779530968835,23.388807580406727,15.785553052788364,Trøndelag
+0.8762852202104873,21.695807900175513,15.829394007404826,Trøndelag
+0.9820552493007314,20.34493948556368,14.15446860736659,Trøndelag
+1.020197782819009,21.46277400495061,17.775800663193444,Trøndelag
+1.052504739975462,20.58272605962876,13.530035221804855,Trøndelag
