+++ testing_minimalist_multiregion/0/dataset_y_test.csv
@@ -1,0 +1,81 @@
+time_period,location,disease_cases
+0.11918459219279837,Finnmark,14.395733540235547
+0.19517200571725712,Finnmark,20.635694354202673
+0.3667322713762387,Finnmark,20.89751365721244
+0.4673444730897248,Finnmark,15.419650897028244
+0.6823396097127724,Finnmark,14.931925134404661
+0.6927907731359264,Finnmark,19.738775415288515
+0.7010283105185549,Finnmark,15.203512297174822
+0.7423361542111717,Finnmark,19.6659646819849
+0.7742298886105831,Finnmark,19.266181806700754
+0.7805727265336266,Finnmark,20.32900634090515
+0.8619971692004202,Finnmark,19.54335217269965
+0.9417656803763995,Finnmark,16.97334748438827
+0.9507722471710189,Finnmark,13.911961664795394
+0.9762185900065296,Finnmark,18.454353291275012
+0.9763827765624724,Finnmark,13.31868915029191
+1.0061822606478055,Finnmark,14.967797713019571
+1.0139730231195048,Finnmark,22.247263856857867
+1.0373265465253905,Finnmark,19.231746831582896
+1.0521538097296175,Finnmark,14.219317478577311
+1.0937288435924677,Finnmark,14.810125345581007
+0.16261640731057428,Troms,21.874032492794882
+0.3272979231388481,Troms,17.897040665056277
+0.4015178682815123,Troms,14.767403146164678
+0.46379089036405624,Troms,16.490375160750663
+0.47636947642974947,Troms,17.041076243209243
+0.5449471863560928,Troms,15.519887305065392
+0.7160920616405975,Troms,19.08284171437311
+0.7443111599469558,Troms,16.083082472410076
+0.744586883443375,Troms,18.147607578591582
+0.7446504990035601,Troms,13.890407334221704
+0.7530497301597526,Troms,23.705668600996464
+0.7942826198787734,Troms,18.969285419570014
+0.85542845325249,Troms,18.818360210853072
+0.9989424592275628,Troms,18.202075564794658
+1.0691066853011042,Troms,19.4039146167379
+1.0870340265402474,Troms,16.963794201602397
+1.0981796177254703,Troms,19.280535539264527
+1.1168133503514053,Troms,15.970740617891389
+1.1175170200631959,Troms,16.345730942301394
+1.2163210157908702,Troms,20.11624853137299
+0.15753632421154296,Nordland,20.262586085642795
+0.16457966770382934,Nordland,15.77111589888969
+0.24737160264135472,Nordland,14.531096653224605
+0.33815055037615555,Nordland,19.768015186481023
+0.3501512300176474,Nordland,13.342163477582217
+0.36660196683035823,Nordland,19.233209968004612
+0.37839435151502365,Nordland,19.551256241452666
+0.39470238184587836,Nordland,20.198639711389053
+0.39568517768248596,Nordland,14.069310784830101
+0.43756720720667525,Nordland,18.84983265956394
+0.4484852917407347,Nordland,19.505932971731404
+0.4586195310589325,Nordland,19.777258821211717
+0.49832623815897614,Nordland,13.652788126398356
+0.5281653937674378,Nordland,16.434430352223856
+0.5458690128724605,Nordland,12.350351005888268
+0.555163923538582,Nordland,18.788751526728753
+0.6192795376719142,Nordland,14.295269699610218
+0.6575427585179277,Nordland,13.431226588124417
+0.669088303052583,Nordland,21.365929471853946
+0.6757272612961174,Nordland,17.76672185431741
+0.07817073492641759,Trøndelag,26.973402154069035
+0.20795968851779276,Trøndelag,23.119995494178976
+0.2368690711721977,Trøndelag,14.029624731092781
+0.4550780725143516,Trøndelag,22.524178530125035
+0.48248558949374765,Trøndelag,23.216688259730244
+0.49078768121970173,Trøndelag,13.368649352940558
+0.5416875533737691,Trøndelag,23.527426486237697
+0.5566197123061257,Trøndelag,13.637024865191577
+0.5573285564903084,Trøndelag,15.04754542946317
+0.5582990440507196,Trøndelag,13.530035221804855
+0.6252029360680726,Trøndelag,19.352199167733513
+0.6649025817969281,Trøndelag,14.449588465450182
+0.7438852362911085,Trøndelag,21.630516822093046
+0.7774361499829475,Trøndelag,13.859246541597509
+0.84653265771923,Trøndelag,18.955596515315918
+0.8478779530968835,Trøndelag,15.829394007404826
+0.8762852202104873,Trøndelag,15.524891611643739
+0.9820552493007314,Trøndelag,14.094962552185436
+1.020197782819009,Trøndelag,16.858038999163814
+1.052504739975462,Trøndelag,13.713870435986205
