+++ testing_minimalist_multiregion/0/predictions.csv
@@ -1,0 +1,81 @@
+time_period,rainfall,mean_temperature,location,sample_0
+0.1191845921927983,19.601666355205992,14.944425883044593,Finnmark,15.140959961998444
+0.1951720057172571,25.19594506858049,22.247263856857867,Finnmark,22.3314714614522
+0.3667322713762387,24.99561860603057,22.649117233487395,Finnmark,22.62370223336836
+0.4673444730897248,20.27702156806342,16.154521780921137,Finnmark,16.28122158810444
+0.6823396097127724,21.03755538387244,15.203619006378062,Finnmark,15.648038480301707
+0.6927907731359264,27.81718966027964,20.13171798636611,Finnmark,21.11193414982194
+0.7010283105185549,18.696245969897276,16.257342437047342,Finnmark,16.044644770962552
+0.7423361542111717,28.83198766574205,19.729266871901643,Finnmark,20.985001820972382
+0.7742298886105831,28.93589496747379,19.12299318081818,Finnmark,20.503744253222482
+0.7805727265336266,20.916073012558197,22.93160430172782,Finnmark,22.027362515457202
+0.8619971692004202,28.36868617503084,19.6659646819849,Finnmark,20.838234724545494
+0.9417656803763996,22.141205066625137,17.760135127740106,Finnmark,17.99123976547559
+0.9507722471710188,20.292529651596443,13.913375806879406,Finnmark,14.42717799058709
+0.9762185900065296,24.80383136966226,19.109201123631117,Finnmark,19.65119238624937
+0.9763827765624724,16.646496949328245,14.103176896055375,Finnmark,13.842276800374998
+1.0061822606478057,22.643022536638924,14.740644763448644,Finnmark,15.591188216989796
+1.0139730231195048,26.286570090287505,24.097532965751466,Finnmark,24.08676628560917
+1.0373265465253905,29.930809750478858,18.7444367084344,Finnmark,20.392565589577234
+1.0521538097296177,23.269415150277563,13.484537130506062,Finnmark,14.677781140291899
+1.0937288435924677,23.50022992091168,14.255474143096537,Finnmark,15.363628470956595
+0.1626164073105742,21.772592619588053,25.0,Troms,24.237350741317286
+0.3272979231388481,21.07023309254854,19.491026136814156,Troms,19.378352499136156
+0.4015178682815123,22.085085705973384,14.717434086823172,Troms,15.325242490251084
+0.4637908903640562,20.725259498132075,17.55666907754366,Troms,17.66269227789125
+0.4763694764297494,21.212295053407075,18.202075564794654,Troms,18.271160267293112
+0.5449471863560928,20.838374046875828,16.124026904753563,Troms,16.427732341749405
+0.7160920616405975,22.03410641712256,20.842984833645325,Troms,20.647476832175713
+0.7443111599469558,21.64318356843154,16.664469170757336,Troms,16.975700673678695
+0.744586883443375,27.936333229069408,17.815700723949284,Troms,18.58652716602015
+0.7446504990035601,17.31933536013001,14.767403146164678,Troms,14.90702232391946
+0.7530497301597526,26.32568860710812,26.214521302793816,Troms,25.73465241209845
+0.7942826198787734,29.591340798555507,18.508147636390657,Troms,19.349050453374712
+0.85542845325249,26.51806264788416,19.16013752412277,Troms,19.61834257752193
+0.9989424592275628,26.735277727861277,18.19306703445013,Troms,18.79835833760059
+1.0691066853011042,25.971239212703377,20.11624853137299,Troms,20.39686532390455
+1.0870340265402474,22.47981899311261,17.62483759970721,Troms,17.89194690193338
+1.0981796177254703,24.711817268344173,20.29484361638377,Troms,20.43017807290489
+1.1168133503514053,22.01082603372266,16.33508258443527,Troms,16.72485984591652
+1.117517020063196,21.41437764823671,17.041076243209243,Troms,17.28105710065488
+1.2163210157908702,23.292336322608342,21.87403249279488,Troms,21.666031735585673
+0.1575363242115429,32.789011788602714,19.551256241452663,Nordland,20.438981412618336
+0.1645796677038293,19.672790588814543,16.881154601431888,Nordland,16.166445507009055
+0.2473716026413547,20.697193970894563,14.802818952469616,Nordland,14.759723385179075
+0.3381505503761555,29.18887310614576,19.842374998866763,Nordland,20.04557253704382
+0.3501512300176474,16.467282701264736,14.295269699610218,Nordland,13.650916515692014
+0.3666019668303582,27.77833186270155,19.4765019420638,Nordland,19.526182797620567
+0.3783943515150236,30.7230368467228,19.08748792846615,Nordland,19.73313002174158
+0.3947023818458783,31.14122696902212,19.89004290251875,Nordland,20.41534437694315
+0.3956851776824859,23.310219256624496,13.371121017404722,Nordland,14.11648470510214
+0.4375672072066752,26.507753139134927,19.27967709553165,Nordland,19.159354293224137
+0.4484852917407347,28.07252053092901,19.768015186481023,Nordland,19.798289833007903
+0.4586195310589325,30.14695898657695,19.561189543098934,Nordland,19.995234673375844
+0.4983262381589761,19.86047333454445,13.744134729871984,Nordland,13.811091162882818
+0.5281653937674378,25.13812593509959,16.20489331135978,Nordland,16.58535005652495
+0.5458690128724605,16.655336672437393,12.791113414080137,Nordland,12.538308449258803
+0.555163923538582,29.43368117280216,18.336279459062418,Nordland,18.94118315123009
+0.6192795376719142,19.695169624337808,14.68846032888861,Nordland,14.501528938301544
+0.6575427585179277,20.275003260047963,13.281886864418103,Nordland,13.530108265104602
+0.669088303052583,32.616362162264764,21.089094918620976,Nordland,21.579854616426292
+0.6757272612961174,28.668732254068463,17.074125903068833,Nordland,17.849961384888736
+0.0781707349264175,27.561399348706136,30.64534542305234,Trøndelag,30.316931453104
+0.2079596885177927,27.40374025848317,25.16327468551458,Trøndelag,25.61505311063197
+0.2368690711721977,21.3712099136426,13.89565494260514,Trøndelag,14.78870981563849
+0.4550780725143516,24.847845807116556,25.0,Trøndelag,24.955568303463107
+0.4824855894937476,29.938401418702345,24.530156721786465,Trøndelag,25.591825625771246
+0.4907876812197017,19.90598492193794,13.326511209723726,Trøndelag,14.005564729350542
+0.5416875533737691,20.211803509054228,27.74294754003143,Trøndelag,26.348146234293573
+0.5566197123061257,18.50793369571012,14.09806251510403,Trøndelag,14.378137463662515
+0.5573285564903084,21.25022074218208,15.329459791782735,Trøndelag,15.985429458518379
+0.5582990440507196,18.807833787002565,13.859246541597509,Trøndelag,14.235769837674319
+0.6252029360680726,23.377716793345005,20.85947350962335,Trøndelag,21.12922336541818
+0.6649025817969281,17.511880137419666,15.52489161164374,Trøndelag,15.390744033679432
+0.7438852362911085,24.346921478807225,23.81695185425238,Trøndelag,23.845822916100552
+0.7774361499829475,19.49740208534876,14.094962552185436,Trøndelag,14.576961445172712
+0.84653265771923,25.66172040636693,19.60238359302318,Trøndelag,20.523443414973226
+0.8478779530968835,23.388807580406727,15.785553052788364,Trøndelag,16.80937756096373
+0.8762852202104873,21.695807900175517,15.829394007404826,Trøndelag,16.50201260501118
+0.9820552493007314,20.34493948556368,14.15446860736659,Trøndelag,14.800216523381678
+1.020197782819009,21.46277400495061,17.775800663193444,Trøndelag,18.112567155333267
+1.052504739975462,20.58272605962876,13.530035221804855,Trøndelag,14.316722588878545
